Measure the balance factor of a binary tree

+++ 14-binary_tree_balance.c
@@ -13,7 +13,7 @@
 
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int left_height, right_height;
+	size_t left_height, right_height;
 
 	if (tree == NULL)
 	{
@@ -21,5 +21,5 @@ int binary_tree_balance(const binary_tree_t *tree)
 	}
 	left_height = binary_tree_height(tree->left);
 	right_height = binary_tree_height(tree->right);
-	return (left_height - right_height);
+	return ((int)(left_height - right_height));
 }
